# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "operationId": "WorkspaceController_listAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseWorkspaceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      },
      "post": {
        "operationId": "WorkspaceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseFullWorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/me": {
      "get": {
        "operationId": "WorkspaceController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResponseWorkspaceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{workspaceId}": {
      "get": {
        "operationId": "WorkspaceController_getWorkspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{workspaceId}/users": {
      "get": {
        "operationId": "WorkspaceController_getWorkspaceUsers",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWorkspaceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{workspaceId}/user/{userId}": {
      "patch": {
        "operationId": "WorkspaceController_addUser",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{workspaceId}/ps": {
      "patch": {
        "operationId": "WorkspaceController_updatePS",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Abhdishba"
                  },
                  "html": {
                    "type": "string",
                    "example": "Abhdishba"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    },
    "/api/workspaces/{id}": {
      "delete": {
        "operationId": "WorkspaceController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Workspace"
        ]
      }
    }
  },
  "info": {
    "title": "Best CRM API",
    "description": "API documentation for Best CRM backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ResponseWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "id"
        ]
      },
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          }
        },
        "required": [
          "name"
        ]
      },
      "ResponseFullWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Test Workspace"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "id",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      }
    }
  }
}